+++ src/Admin/AdminInterface.php
@@ -143,6 +143,8 @@ interface AdminInterface extends AccessRegistryInterface, FieldDescriptionRegist
     public function getForm();
 
     /**
+     * NEXT MAJOR: Remove the throws tag.
+     *
      * @throws \RuntimeException if no request is set
      *
      * @return Request
@@ -355,6 +357,8 @@ interface AdminInterface extends AccessRegistryInterface, FieldDescriptionRegist
     public function setSubject($subject);
 
     /**
+     * NEXT MAJOR: return object.
+     *
      * @return object|null
      */
     public function getSubject();
@@ -450,7 +454,7 @@ interface AdminInterface extends AccessRegistryInterface, FieldDescriptionRegist
     public function createObjectSecurity($object);
 
     /**
-     * @return AdminInterface|null
+     * @return AdminInterface|null NEXT_MAJOR: return AdminInterface
      */
     public function getParent();